# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BaraGo API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /residents:
    get:
      operationId: listResidents
      tags: [residents]
      summary: List all residents (admin only)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: verified
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of residents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResidentWithUser"
  /residents/{id}:
    get:
      operationId: getResident
      tags: [residents]
      summary: Get a resident by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentWithUser"
        "404":
          description: Not found

    patch:
      operationId: updateResident
      tags: [residents]
      summary: Update resident details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResidentUpdate"
      responses:
        "200":
          description: Updated resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentWithUser"
  /residents/{id}/verify:
    patch:
      operationId: verifyResident
      tags: [residents]
      summary: Verify or unverify a resident (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentWithUser"
  /residents/{id}/disable:
    patch:
      operationId: disableResident
      tags: [residents]
      summary: Disable or enable a resident account (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisableInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentWithUser"

  # Appointments
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List health schedules
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HealthSchedule"

    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a health schedule (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthSchedule"
  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      summary: Get health schedule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthSchedule"

    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthSchedule"

    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # Notifications
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List generated reports (admin only)
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a report (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, email, role, status, createdAt]
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [resident, admin, health_worker]
        status:
          type: string
          enum: [active, disabled]
        createdAt:
          type: string
...
    ResidentWithUser:
      type: object
      required: [id, userId, address, contactNumber, verified, createdAt, user]
      properties:
        id:
          type: string
        userId:
          type: string
        birthdate:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: string
        purok:
          type: ["string", "null"]
        contactNumber:
          type: string
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactNumber:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
    ResidentUpdate:
      type: object
      properties:
        birthdate:
          type: string
        age:
          type: integer
        gender:
          type: string
        address:
          type: string
        purok:
          type: string
        contactNumber:
          type: string
        emergencyContactName:
          type: string
        emergencyContactNumber:
          type: string
    VerifyInput:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean
    DisableInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, disabled]

    # Appointments
...
    HealthSchedule:
      type: object
      required: [id, scheduleDate, startTime, endTime, slotLimit, currentSlots, status, createdAt]
      properties:
        id:
          type: string
        scheduleDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        slotLimit:
          type: integer
        currentSlots:
          type: integer
        assignedStaff:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, closed, cancelled]
        createdAt:
          type: string
    ScheduleInput:
      type: object
      required: [scheduleDate, startTime, endTime, slotLimit]
      properties:
        scheduleDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        slotLimit:
          type: integer
        assignedStaff:
          type: string
        status:
          type: string
          enum: [open, closed, cancelled]
    ScheduleUpdate:
      type: object
      properties:
        scheduleDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        slotLimit:
          type: integer
        assignedStaff:
          type: string
        status:
          type: string
          enum: [open, closed, cancelled]

    # Notifications
...
    Report:
      type: object
      required: [id, reportType, createdAt]
      properties:
        id:
          type: string
        generatedBy:
          type: ["string", "null"]
        reportType:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ReportInput:
      type: object
      required: [reportType]
      properties:
        reportType:
          type: string
          enum: [appointments, ambulance, residents, completed_checkups, monthly_summary]
        startDate:
          type: string
        endDate:
          type: string